Calculate random virtual KASLR offset
The implementation in linux/arch/x86/boot/compressed/kaslr.c uses
KERNEL_IMAGE_SIZE, CONFIG_PHYSICAL_START, and CONFIG_PHYSICAL_ALIGN
to calculate the number of CONFIG_PHYSICAL_ALIGN sized slots
that can hold the kernel's actual size between the addresses
CONFIG_PHYSICAL_START and KERNEL_IMAGE_SIZE. KERNEL_IMAGE_SIZE
the maximum virtual address for the kernel image and is hardcoded
to be 1GiB when Linux is compiled with KASLR support
(CONFIG_RANDOMIZE_BASE). CONFIG_PHYSICAL_START and CONFIG_PHYSICAL_ALIGN
are config options that are not directly available from the kernel ELF.
To determine what they are, this implementation uses the smallest
physical address of all the loadable semgments as CONFIG_PHYSICAL_START
and the maximum alignment of each loadable segment as
CONFIG_PHYSICAL_ALIGN. With all of this information, we can calculate
the number of slots the kernel will fit in, and pick one at random
the same way it is done in the Linux implementation.

Signed-off-by: Ben Holmes <bholmes@vassar.edu>